Fix accounts to make all unit tests green.

+++ dto/src/main/java/capitalone/interview/dto/Transaction.java
@@ -59,7 +59,7 @@ public class Transaction {
         this.previousTransactionId = previous_transaction_id;
     }
 
-    public int getAmount() {
+    public long getAmount() {
         return amount;
     }
